requirements.txt
# API
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Config
pydantic-settings==2.3.4

# Task queue
celery[redis]==5.4.0

# Discord bot
discord.py==2.3.2

# Utils
python-multipart==0.0.9

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
